UI omaksi luokakseen
